re grpc: share Google API protos
The remote-execution proto crate currently owns common Google API,
Bytestream, long-running operation, and status definitions.

Move those definitions into a shared crate because BES needs the same
types and compiling duplicate packages would create incompatible Rust
types. Keep re_grpc_proto re-exporting them so existing users retain
their API.
