+++ Cargo.toml
@@ -148,6 +148,7 @@ members = [
   "pagable",
   "pagable_derive",
   "pagable_storage",
+  "remote_execution/oss/google_grpc_proto",
   "remote_execution/oss/re_grpc",
   "remote_execution/oss/re_grpc_proto",
   "shed/lock_free_hashtable",
@@ -336,6 +337,7 @@ hostname = "0.4.2"
 http = "1.4.2"
 http-body = "1.1.0"
 http-body-util = "0.1.4"
+google_grpc_proto = { path = "remote_execution/oss/google_grpc_proto" }
 httparse = "1.7.1"
 httptest = "0.16.4"
 humantime = "2.4.0"
