+++ remote_execution/oss/google_grpc_proto/BUCK
@@ -1,19 +1,22 @@
 load("@fbcode//buck2:proto_defs.bzl", "rust_protobuf_library")
+load("@fbcode_macros//build_defs:native_rules.bzl", "buck_filegroup")
 
 oncall("build_infra")
 
 rust_protobuf_library(
-    name = "re_grpc_proto",
+    name = "google_grpc_proto",
     srcs = glob(["src/**/*.rs"]),
     build_script = "build.rs",
     protos = glob(["proto/**/*.proto"]),
-    build_env = {
-        "GOOGLE_PROTO_ROOT": "$(location //remote_execution/oss/google_grpc_proto:google_grpc_proto_sources)",
-    },
     deps = [
         "fbsource//third-party/rust:prost-types",
         "fbsource//third-party/rust:serde",
         "//buck2/app/buck2_data:buck2_data",
-        "//buck2/remote_execution/oss/google_grpc_proto:google_grpc_proto",
     ],
+)
+
+buck_filegroup(
+    name = "google_grpc_proto_sources",
+    srcs = glob(["proto/**/*.proto"]),
+    visibility = ["PUBLIC"],
 )
