+++ remote_execution/oss/google_grpc_proto/proto/google/api/launch_stage.proto
