+++ remote_execution/oss/re_grpc_proto/Cargo.toml
@@ -9,6 +9,7 @@ license.workspace = true
 
 [dependencies]
 buck2_data.workspace = true
+google_grpc_proto.workspace = true
 prost.workspace = true
 prost-types.workspace = true
 serde.workspace = true
