+++ remote_execution/oss/re_grpc_proto/build.rs
@@ -14,40 +14,33 @@ fn main() -> io::Result<()> {
     let proto_files = &[
         "proto/build/bazel/remote/execution/v2/remote_execution.proto",
         "proto/build/bazel/semver/semver.proto",
-        "proto/google/api/annotations.proto",
-        "proto/google/api/client.proto",
-        "proto/google/api/http.proto",
-        "proto/google/bytestream/bytestream.proto",
-        "proto/google/longrunning/operations.proto",
-        "proto/google/rpc/code.proto",
-        "proto/google/rpc/status.proto",
     ];
+
+    let google_proto_root = std::env::var("GOOGLE_PROTO_ROOT").ok();
+    let google_proto_root = google_proto_root
+        .as_deref()
+        .map(|root| format!("{root}/proto"));
+    let mut includes = vec!["./proto/", "../google_grpc_proto/proto/"];
+    if let Some(include) = google_proto_root.as_deref() {
+        includes.push(include);
+    }
 
     let builder = buck2_protoc_dev::configure();
     unsafe { builder.setup_protoc() }
         .type_attribute(".", "#[derive(::serde::Serialize, ::serde::Deserialize)]")
+        .extern_path(".google.api", "::google_grpc_proto::google::api")
+        .extern_path(".google.bytestream", "::google_grpc_proto::google::bytestream")
+        .extern_path(
+            ".google.longrunning",
+            "::google_grpc_proto::google::longrunning",
+        )
+        .extern_path(".google.rpc", "::google_grpc_proto::google::rpc")
         .field_attribute(
             "build.bazel.remote.execution.v2.Action.timeout",
             "#[serde(with = \"::buck2_data::serialize_duration_as_micros\")]",
         )
         .field_attribute(
             "build.bazel.remote.execution.v2.ExecutedActionMetadata.virtual_execution_duration",
-            "#[serde(with = \"::buck2_data::serialize_duration_as_micros\")]",
-        )
-        .field_attribute(
-            "google.longrunning.WaitOperationRequest.timeout",
-            "#[serde(with = \"::buck2_data::serialize_duration_as_micros\")]",
-        )
-        .field_attribute(
-            "google.api.MethodSettings.LongRunning.initial_poll_delay",
-            "#[serde(with = \"::buck2_data::serialize_duration_as_micros\")]",
-        )
-        .field_attribute(
-            "google.api.MethodSettings.LongRunning.max_poll_delay",
-            "#[serde(with = \"::buck2_data::serialize_duration_as_micros\")]",
-        )
-        .field_attribute(
-            "google.api.MethodSettings.LongRunning.total_poll_timeout",
             "#[serde(with = \"::buck2_data::serialize_duration_as_micros\")]",
         )
         .field_attribute(
@@ -92,19 +85,7 @@ fn main() -> io::Result<()> {
         )
         .field_attribute(
             "build.bazel.remote.execution.v2.ExecutedActionMetadata.auxiliary_metadata",
-            "#[serde(with = \"crate::serialize_vec_any\")]",
+            "#[serde(with = \"::google_grpc_proto::serialize_vec_any\")]",
         )
-        .field_attribute(
-            "google.longrunning.Operation.metadata",
-            "#[serde(with = \"crate::serialize_option_any\")]",
-        )
-        .field_attribute(
-            "google.longrunning.Operation.result.response",
-            "#[serde(with = \"crate::serialize_any\")]",
-        )
-        .field_attribute(
-            "google.rpc.Status.details",
-            "#[serde(with = \"crate::serialize_vec_any\")]",
-        )
-        .compile(proto_files, &["./proto/"])
+        .compile(proto_files, &includes)
 }
