+++ remote_execution/oss/re_grpc_proto/src/lib.rs
@@ -8,21 +8,7 @@
  * above-listed licenses.
  */
 
-#[allow(clippy::doc_lazy_continuation)]
-pub mod google {
-    pub mod api {
-        tonic::include_proto!("google.api");
-    }
-    pub mod bytestream {
-        tonic::include_proto!("google.bytestream");
-    }
-    pub mod longrunning {
-        tonic::include_proto!("google.longrunning");
-    }
-    pub mod rpc {
-        tonic::include_proto!("google.rpc");
-    }
-}
+pub use google_grpc_proto::google;
 
 #[allow(clippy::doc_lazy_continuation)]
 pub mod build {
@@ -37,90 +23,5 @@ pub mod build {
                 }
             }
         }
-    }
-}
-
-pub mod serialize_vec_any {
-    use serde::Deserialize;
-    use serde::Deserializer;
-    use serde::Serialize;
-    use serde::Serializer;
-
-    pub fn serialize<S>(value: &[::prost_types::Any], serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let d: Vec<(String, Vec<u8>)> = value
-            .iter()
-            .map(|v| (v.type_url.clone(), v.value.clone()))
-            .collect();
-        d.serialize(serializer)
-    }
-
-    pub fn deserialize<'de, D>(deserializer: D) -> Result<Vec<::prost_types::Any>, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let d: Vec<::prost_types::Any> = Vec::deserialize(deserializer)?
-            .into_iter()
-            .map(|(type_url, value)| ::prost_types::Any { type_url, value })
-            .collect();
-        Ok(d)
-    }
-}
-
-pub mod serialize_option_any {
-    use serde::Deserialize;
-    use serde::Deserializer;
-    use serde::Serialize;
-    use serde::Serializer;
-
-    pub fn serialize<S>(
-        value: &Option<::prost_types::Any>,
-        serializer: S,
-    ) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let d = value
-            .as_ref()
-            .map(|v| (v.type_url.clone(), v.value.clone()));
-        d.serialize(serializer)
-    }
-
-    pub fn deserialize<'de, D>(deserializer: D) -> Result<Option<::prost_types::Any>, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let d = Option::<(String, Vec<u8>)>::deserialize(deserializer)?
-            .map(|(type_url, value)| ::prost_types::Any { type_url, value });
-        Ok(d)
-    }
-}
-
-pub mod serialize_any {
-    use serde::Deserialize;
-    use serde::Deserializer;
-    use serde::Serialize;
-    use serde::Serializer;
-
-    pub fn serialize<S>(value: &::prost_types::Any, serializer: S) -> Result<S::Ok, S::Error>
-    where
-        S: Serializer,
-    {
-        let d = (value.type_url.clone(), value.value.clone());
-        d.serialize(serializer)
-    }
-
-    pub fn deserialize<'de, D>(deserializer: D) -> Result<::prost_types::Any, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let d = <(String, Vec<u8>)>::deserialize(deserializer)?;
-        let d = ::prost_types::Any {
-            type_url: d.0,
-            value: d.1,
-        };
-        Ok(d)
     }
 }
